+++ DrillingRig.ConfigApp/AppControl/AinSettingsWrite/IAinSettingsPart.cs
@@ -6,7 +6,7 @@ namespace DrillingRig.ConfigApp.AppControl.AinSettingsWrite {
 		double? KpW { get; }
 		double? KiW { get; }
 
-		short? FiNom { get; }
+		double? FiNom { get; }
 		short? Imax { get; }
 
 		short? UdcMax { get; }
@@ -16,10 +16,10 @@ namespace DrillingRig.ConfigApp.AppControl.AinSettingsWrite {
 		double? Fmax { get; }
 
 		double? DflLim { get; }
-		short? FlMinMin { get; }
+		double? FlMinMin { get; }
 
 		short? IoutMax { get; }
-		short? FiMin { get; }
+		double? FiMin { get; }
 		short? DacCh { get; }
 
 		short? Imcw { get; }
